Downloads/RDT 6/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# AI and ML
openai==1.3.7
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.1

# Vector Database
chromadb==0.4.18
pinecone-client==2.2.4
azure-search-documents==11.4.0

# Database
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Search and Processing
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
nltk==3.8.1
spacy==3.7.2

# Utilities
python-dotenv==1.0.0
pydantic-settings==2.1.0
loguru==0.7.2
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
deepeval==0.20.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# HTTP and API
httpx==0.25.2
requests==2.31.0
tiktoken==0.5.2

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Configuration
pyyaml==6.0.1
toml==0.10.2

# Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-azure-monitor==1.0.0

# Authentication
PyJWT==2.8.0
msal==1.24.1

# Frontend
vue==3.3.8
@microsoft/teamsfx==2.0.0 